♻️(common): rename response-protocole.ts to response-protocol.ts
Fix filename typo 'protocole' -> 'protocol' in
packages/common/src/middleware/. Update all imports
and mock paths to reflect the new filename.

Closes #109
