+++ packages/common/src/middleware/response-protocol.ts
